changes of formating

+++ codec_util/main/codec_util.cpp
@@ -73,7 +73,7 @@ codec_data::reader &codec_data::reader::operator>> (unsigned long &d) {
     memcpy (&d, dt, sizeof (unsigned long));
     if (readed_bit) {
       d >>= readed_bit;
-      d |= (unsigned long(*(dt + sizeof (unsigned long))) >> (CHAR_BIT - readed_bit)) << ((sizeof (unsigned long) - 1) << 3);
+      d |= (unsigned long (*(dt + sizeof (unsigned long))) >> (CHAR_BIT - readed_bit)) << ((sizeof (unsigned long) - 1) << 3);
     }
     readed_byte += sizeof (unsigned long);
   }
